+++ src/main/java/io/github/arkosammy12/jemu/frontend/oldui/MainWindow.java
@@ -1,13 +1,13 @@
-package io.github.arkosammy12.jemu.frontend.ui;
+package io.github.arkosammy12.jemu.frontend.oldui;
 
 import io.github.arkosammy12.jemu.application.Jemu;
 import io.github.arkosammy12.jemu.application.Main;
 import io.github.arkosammy12.jemu.application.io.DataManager;
 import io.github.arkosammy12.jemu.application.io.initializers.EmulatorInitializer;
 import io.github.arkosammy12.jemu.application.io.initializers.EmulatorInitializerConsumer;
-import io.github.arkosammy12.jemu.frontend.ui.debugger.DebuggerPanel;
-import io.github.arkosammy12.jemu.frontend.ui.util.ToggleableSplitPane;
-import io.github.arkosammy12.jemu.frontend.ui.util.WindowTitleManager;
+import io.github.arkosammy12.jemu.frontend.oldui.debugger.DebuggerPanel;
+import io.github.arkosammy12.jemu.frontend.oldui.util.ToggleableSplitPane;
+import io.github.arkosammy12.jemu.frontend.oldui.util.WindowTitleManager;
 import net.miginfocom.layout.AC;
 import net.miginfocom.layout.CC;
 import net.miginfocom.layout.LC;
