+++ src/main/java/io/github/arkosammy12/jemu/frontend/oldui/SettingsBar.java
@@ -1,10 +1,10 @@
-package io.github.arkosammy12.jemu.frontend.ui;
+package io.github.arkosammy12.jemu.frontend.oldui;
 
 import io.github.arkosammy12.jemu.application.Jemu;
 import io.github.arkosammy12.jemu.application.io.initializers.EmulatorInitializer;
 import io.github.arkosammy12.jemu.application.util.System;
 import io.github.arkosammy12.jemu.application.util.KeyboardLayout;
-import io.github.arkosammy12.jemu.frontend.ui.menus.*;
+import io.github.arkosammy12.jemu.frontend.oldui.menus.*;
 
 import javax.swing.*;
 import java.nio.file.Path;
