+++ src/main/java/io/github/arkosammy12/jemu/frontend/oldui/menus/DebugMenu.java
@@ -1,11 +1,11 @@
-package io.github.arkosammy12.jemu.frontend.ui.menus;
+package io.github.arkosammy12.jemu.frontend.oldui.menus;
 
 import io.github.arkosammy12.jemu.application.Jemu;
 import io.github.arkosammy12.jemu.application.io.DataManager;
 import io.github.arkosammy12.jemu.application.io.initializers.ApplicationInitializer;
 import io.github.arkosammy12.jemu.application.io.initializers.EmulatorInitializer;
 import io.github.arkosammy12.jemu.application.io.initializers.EmulatorInitializerConsumer;
-import io.github.arkosammy12.jemu.frontend.ui.MainWindow;
+import io.github.arkosammy12.jemu.frontend.oldui.MainWindow;
 
 import javax.swing.*;
 import java.awt.event.KeyEvent;
