+++ src/main/java/io/github/arkosammy12/jemu/frontend/oldui/menus/SettingsMenu.java
@@ -1,4 +1,4 @@
-package io.github.arkosammy12.jemu.frontend.ui.menus;
+package io.github.arkosammy12.jemu.frontend.oldui.menus;
 
 import io.github.arkosammy12.jemu.application.Jemu;
 import io.github.arkosammy12.jemu.application.io.DataManager;
@@ -6,8 +6,8 @@ import io.github.arkosammy12.jemu.application.io.Serializable;
 import io.github.arkosammy12.jemu.application.io.initializers.ApplicationInitializer;
 import io.github.arkosammy12.jemu.application.io.initializers.EmulatorInitializer;
 import io.github.arkosammy12.jemu.application.io.initializers.EmulatorInitializerConsumer;
-import io.github.arkosammy12.jemu.frontend.ui.MainWindow;
-import io.github.arkosammy12.jemu.frontend.ui.util.EnumMenu;
+import io.github.arkosammy12.jemu.frontend.oldui.MainWindow;
+import io.github.arkosammy12.jemu.frontend.oldui.util.EnumMenu;
 import io.github.arkosammy12.jemu.application.util.KeyboardLayout;
 
 import javax.swing.*;
