+++ src/main/java/io/github/arkosammy12/jemu/frontend/oldui/util/EnumMenu.java
@@ -1,4 +1,4 @@
-package io.github.arkosammy12.jemu.frontend.ui.util;
+package io.github.arkosammy12.jemu.frontend.oldui.util;
 
 import io.github.arkosammy12.jemu.application.util.DisplayNameProvider;
